+++ app/controllers/dashboard_controller.rb
@@ -9,9 +9,6 @@ class DashboardController < ApplicationController
     @recent_donations = @donations.recent
 
     distributions = current_organization.distributions.includes(:partner).during(helpers.selected_range)
-    @recent_distributions = distributions.recent
-
-    @itemized_donation_data = DonationItemizedBreakdownService.new(organization: current_organization, donation_ids: @donations.pluck(:id)).fetch
     @itemized_distribution_data = DistributionItemizedBreakdownService.new(organization: current_organization, distribution_ids: distributions.pluck(:id)).fetch
 
     @total_inventory = current_organization.total_inventory
