+++ app/views/dashboard/index.html.erb
@@ -143,30 +143,6 @@
 
         <%= render(
           "shared/card",
-          id: "distributions",
-          gradient: "info",
-          title: "Distributions",
-          subtitle: @selected_date_range,
-          footer: link_to("See more...", distributions_path),
-          footer_options: { class: "text-center" },
-        ) do %>
-          <%= new_button_to new_distribution_path, {text: "New Distribution"} %>
-
-          <h3 class="text-center">
-            <span class="total_distributed">
-              <%= total_distributed %>
-            </span>
-            items distributed <%= @selected_date_range_label %>
-          </h3>
-          <h4 class="text-center">(<%= future_distributed %> items scheduled for future distribution)</h4>
-          <div class="box-body">
-            <h4>Recent distributions</h4>
-            <%= render partial: "distribution", collection: @recent_distributions, as: :distribution %>
-          </div>
-        <% end %>
-
-        <%= render(
-          "shared/card",
           gradient: "warning",
           title: "Activity",
           subtitle: @selected_date_range,
