+++ app/views/distributions_summary/_distribution.html.erb
