+++ app/views/layouts/_lte_sidebar.html.erb
@@ -170,8 +170,8 @@
       </li>
     </ul>
   </li>
-  <li class="nav-item has-treeview <%= menu_open?(['audits', 'reports', "purchases_summary"]) %>">
-    <a href="#" class="nav-link <%= active_class(['audits', 'reports', "purchases_summary"]) %>">
+  <li class="nav-item has-treeview <%= menu_open?(['audits', 'reports/annual_reports', "distributions_by_county", 'reports/itemized_donations', 'reports/itemized_distributions', "purchases_summary", "distributions_summary"]) %>">
+    <a href="#" class="nav-link <%= active_class(['audits', 'reports/annual_reports', "distributions_by_county", 'reports/itemized_donations', 'reports/itemized_distributions', "purchases_summary", "distributions_summary"]) %>">
       <i class="nav-icon fa fa-tasks"></i>
       <p>Reporting & Auditing
         <i class="fas fa-angle-left right"></i>
@@ -185,29 +185,34 @@
           <% end %>
         </li>
       <% end %>
-      <li class="nav-item <%= active_class(['reports']) %>">
-        <%= link_to(reports_annual_reports_path(organization_id: current_user.organization), class: "nav-link #{active_class(['reports'])}") do %>
+      <li class="nav-item <%= active_class(['reports/annual_reports']) %>">
+        <%= link_to(reports_annual_reports_path(organization_id: current_user.organization), class: "nav-link #{active_class(['reports/annual_reports'])}") do %>
           <i class="nav-icon fa fa-circle-o"></i> Annual Survey
         <% end %>
       </li>
-      <li class="nav-item <%= active_class(['reports']) %>">
-        <%= link_to(distributions_by_county_report_path(filters: { date_range: date_range_params }), class: "nav-link #{active_class(['reports'])}") do %>
+      <li class="nav-item <%= active_class(['distributions_by_county']) %>">
+        <%= link_to(distributions_by_county_report_path(filters: { date_range: date_range_params }), class: "nav-link #{active_class(['distributions_by_county'])}") do %>
           <i class="nav-icon fa fa-circle-o"></i> Distributions by County
         <% end %>
       </li>
-      <li class="nav-item <%= active_class(['reports']) %>">
-        <%= link_to(reports_itemized_donations_path(organization_id: current_user.organization), class: "nav-link #{active_class(['reports'])}") do %>
+      <li class="nav-item <%= active_class(['reports/itemized_donations']) %>">
+        <%= link_to(reports_itemized_donations_path(organization_id: current_user.organization), class: "nav-link #{active_class(['reports/itemized_donations'])}") do %>
           <i class="nav-icon fa fa-circle-o"></i> Itemized Donations
         <% end %>
       </li>
-      <li class="nav-item <%= active_class(['reports']) %>">
-        <%= link_to(reports_itemized_distributions_path(organization_id: current_user.organization), class: "nav-link #{active_class(['reports'])}") do %>
+      <li class="nav-item <%= active_class(['reports/itemized_distributions']) %>">
+        <%= link_to(reports_itemized_distributions_path(organization_id: current_user.organization), class: "nav-link #{active_class(['reports/itemized_distributions'])}") do %>
           <i class="nav-icon fa fa-circle-o"></i> Itemized Distributions
         <% end %>
       </li>
       <li class="nav-item <%= active_class(['purchases_summary']) %>">
         <%= link_to(purchases_summary_path(organization_id: current_user.organization), class: "nav-link #{"active" if current_page?(purchases_summary_path)}") do %>
           <i class="nav-icon fa fa-circle-o"></i> Purchases Summary
+        <% end %>
+      </li>
+      <li class="nav-item <%= active_class(['distributions_summary']) %>">
+        <%= link_to(distributions_summary_path(organization_id: current_user.organization), class: "nav-link #{"active" if current_page?(distributions_summary_path)}") do %>
+          <i class="nav-icon fa fa-circle-o"></i> Distributions Summary
         <% end %>
       </li>
     </ul>
