+++ spec/support/pages/organization_dashboard_page.rb
@@ -6,12 +6,6 @@ class OrganizationDashboardPage < OrganizationPage
     "dashboard"
   end
 
-  def create_new_distribution
-    within distributions_section do
-      click_link "New Distribution"
-    end
-  end
-
   def create_new_donation
     within donations_section do
       click_link "New Donation"
@@ -62,10 +56,6 @@ class OrganizationDashboardPage < OrganizationPage
     has_selector? product_drives_selector
   end
 
-  def has_distributions_section?
-    has_selector? distributions_section_selector
-  end
-
   def has_getting_started_guide?
     has_selector? "#getting-started-guide"
   end
@@ -106,12 +96,6 @@ class OrganizationDashboardPage < OrganizationPage
     end
   end
 
-  def recent_distribution_links
-    within distributions_section do
-      all(".distribution a").map(&:text)
-    end
-  end
-
   def recent_donation_links
     within donations_section do
       all(".donation a").map(&:text)
@@ -136,12 +120,6 @@ class OrganizationDashboardPage < OrganizationPage
 
   def summary_section
     find "#summary"
-  end
-
-  def total_distributed
-    within distributions_section do
-      parse_formatted_integer find(".total_distributed").text
-    end
   end
 
   def total_donations
@@ -188,14 +166,6 @@ class OrganizationDashboardPage < OrganizationPage
 
   def product_drives_selector
     "#product_drives"
-  end
-
-  def distributions_section_selector
-    "#distributions"
-  end
-
-  def distributions_section
-    find distributions_section_selector
   end
 
   def donations_section
